refactor: Update server initialization logic to check database type before removal

+++ src/main.py
@@ -94,12 +94,11 @@ def server_init():
     Initializes the server by checking if the database is already set up.
     If not, it creates the necessary tables and a default admin user.
     """
-    if os.path.exists(ROOT_ABSPATH / "app.db"):
+    if (
+        os.path.exists(ROOT_ABSPATH / "app.db")
+        and global_config["database"]["type"] == "sqlite"
+    ):
         os.remove(ROOT_ABSPATH / "app.db")
-    if os.path.exists(ROOT_ABSPATH / "ssl_cert.pem"):
-        os.remove(ROOT_ABSPATH / "ssl_cert.pem")
-    if os.path.exists(ROOT_ABSPATH / "ssl_key.pem"):
-        os.remove(ROOT_ABSPATH / "ssl_key.pem")
 
     # Create database tables before inserting data
     Base.metadata.create_all(engine)
